@@ -5,16 +5,14 @@ namespace LiteNetLib
 {
     internal abstract class BaseChannel
     {
-        protected readonly NetPeer Peer;
+        protected readonly LiteNetPeer Peer;
         protected readonly Queue<NetPacket> OutgoingQueue = new Queue<NetPacket>(NetConstants.DefaultWindowSize);
         private int _isAddedToPeerChannelSendQueue;
 
         public int PacketsInQueue => OutgoingQueue.Count;
 
-        protected BaseChannel(NetPeer peer)
-        {
+        protected BaseChannel(LiteNetPeer peer) =>
             Peer = peer;
-        }
 
         public void AddToQueue(NetPacket packet)
         {
@@ -28,9 +26,7 @@ namespace LiteNetLib
         protected void AddToPeerChannelSendQueue()
         {
             if (Interlocked.CompareExchange(ref _isAddedToPeerChannelSendQueue, 1, 0) == 0)
-            {
                 Peer.AddToReliableChannelSendQueue(this);
-            }
         }
 
         public bool SendAndCheckQueue()
@@ -42,7 +38,8 @@ namespace LiteNetLib
             return hasPacketsToSend;
         }
 
-        protected abstract bool SendNextPackets();
+        public abstract bool SendNextPackets();
+
         public abstract bool ProcessPacket(NetPacket packet);
     }
 }
