@@ -18,13 +18,34 @@ namespace LiteNetLib
     /// </summary>
     public enum DisconnectReason
     {
+        /// <summary>
+        /// Connection to host failed
+        /// </summary>
         ConnectionFailed,
+
+        /// <summary>
+        /// Timeout
+        /// </summary>
         Timeout,
+
         HostUnreachable,
         NetworkUnreachable,
+
+        /// <summary>
+        /// Remote host disconnected peer
+        /// </summary>
         RemoteConnectionClose,
+
+        /// <summary>
+        /// Disconnect called locally
+        /// </summary>
         DisconnectPeerCalled,
+
+        /// <summary>
+        /// Connection rejected by remote host
+        /// </summary>
         ConnectionRejected,
+
         InvalidProtocol,
         UnknownHost,
         Reconnect,
@@ -53,6 +74,9 @@ namespace LiteNetLib
         public NetPacketReader AdditionalData;
     }
 
+    /// <summary>
+    /// Interface for implementing own INetEventListener. This is a bit faster than use EventBasedListener
+    /// </summary>
     public interface INetEventListener
     {
         /// <summary>
@@ -104,38 +128,101 @@ namespace LiteNetLib
         /// </summary>
         /// <param name="request">Request information (EndPoint, internal id, additional data)</param>
         void OnConnectionRequest(ConnectionRequest request);
-    }
 
-    public interface IDeliveryEventListener
-    {
         /// <summary>
         /// On reliable message delivered
         /// </summary>
         /// <param name="peer"></param>
         /// <param name="userData"></param>
-        void OnMessageDelivered(NetPeer peer, object userData);
-    }
+        void OnMessageDelivered(NetPeer peer, object userData) { }
 
-    public interface INtpEventListener
-    {
         /// <summary>
         /// Ntp response
         /// </summary>
         /// <param name="packet"></param>
-        void OnNtpResponse(NtpPacket packet);
-    }
+        void OnNtpResponse(NtpPacket packet) { }
 
-    public interface IPeerAddressChangedListener
-    {
         /// <summary>
         /// Called when peer address changed (when AllowPeerAddressChange is enabled)
         /// </summary>
         /// <param name="peer">Peer that changed address (with new address)</param>
         /// <param name="previousAddress">previous IP</param>
-        void OnPeerAddressChanged(NetPeer peer, IPEndPoint previousAddress);
+        void OnPeerAddressChanged(NetPeer peer, IPEndPoint previousAddress) { }
     }
 
-    public class EventBasedNetListener : INetEventListener, IDeliveryEventListener, INtpEventListener, IPeerAddressChangedListener
+    /// <summary>
+    /// Interface for implementing own ILiteNetEventListener. This is a bit faster than use EventBasedListener
+    /// </summary>
+    public interface ILiteNetEventListener
+    {
+        /// <summary>
+        /// New remote peer connected to host, or client connected to remote host
+        /// </summary>
+        /// <param name="peer">Connected peer object</param>
+        void OnPeerConnected(LiteNetPeer peer);
+
+        /// <summary>
+        /// Peer disconnected
+        /// </summary>
+        /// <param name="peer">disconnected peer</param>
+        /// <param name="disconnectInfo">additional info about reason, errorCode or data received with disconnect message</param>
+        void OnPeerDisconnected(LiteNetPeer peer, DisconnectInfo disconnectInfo);
+
+        /// <summary>
+        /// Network error (on send or receive)
+        /// </summary>
+        /// <param name="endPoint">From endPoint (can be null)</param>
+        /// <param name="socketError">Socket error</param>
+        void OnNetworkError(IPEndPoint endPoint, SocketError socketError) { }
+
+        /// <summary>
+        /// Received some data
+        /// </summary>
+        /// <param name="peer">From peer</param>
+        /// <param name="reader">DataReader containing all received data</param>
+        /// <param name="deliveryMethod">Type of received packet</param>
+        void OnNetworkReceive(LiteNetPeer peer, NetPacketReader reader, DeliveryMethod deliveryMethod);
+
+        /// <summary>
+        /// Received unconnected message
+        /// </summary>
+        /// <param name="remoteEndPoint">From address (IP and Port)</param>
+        /// <param name="reader">Message data</param>
+        /// <param name="messageType">Message type (simple, discovery request or response)</param>
+        void OnNetworkReceiveUnconnected(IPEndPoint remoteEndPoint, NetPacketReader reader, UnconnectedMessageType messageType) { }
+
+        /// <summary>
+        /// Latency information updated
+        /// </summary>
+        /// <param name="peer">Peer with updated latency</param>
+        /// <param name="latency">latency value in milliseconds</param>
+        void OnNetworkLatencyUpdate(LiteNetPeer peer, int latency) { }
+
+        /// <summary>
+        /// On peer connection requested
+        /// </summary>
+        /// <param name="request">Request information (EndPoint, internal id, additional data)</param>
+        void OnConnectionRequest(ConnectionRequest request);
+
+        /// <summary>
+        /// Called when peer address changed (when AllowPeerAddressChange is enabled)
+        /// </summary>
+        /// <param name="peer">Peer that changed address (with new address)</param>
+        /// <param name="previousAddress">previous IP</param>
+        void OnPeerAddressChanged(LiteNetPeer peer, IPEndPoint previousAddress) { }
+
+        /// <summary>
+        /// On reliable message delivered
+        /// </summary>
+        /// <param name="peer"></param>
+        /// <param name="userData"></param>
+        void OnMessageDelivered(LiteNetPeer peer, object userData) { }
+    }
+
+    /// <summary>
+    /// Simple event based listener for simple setups and benchmarks
+    /// </summary>
+    public class EventBasedNetListener : INetEventListener
     {
         public delegate void OnPeerConnected(NetPeer peer);
         public delegate void OnPeerDisconnected(NetPeer peer, DisconnectInfo disconnectInfo);
@@ -159,114 +246,108 @@ namespace LiteNetLib
         public event OnNtpResponseEvent NtpResponseEvent;
         public event OnPeerAddressChangedEvent PeerAddressChangedEvent;
 
-        public void ClearPeerConnectedEvent()
-        {
-            PeerConnectedEvent = null;
-        }
+        public void ClearPeerConnectedEvent() =>  PeerConnectedEvent = null;
+        public void ClearPeerDisconnectedEvent() => PeerDisconnectedEvent = null;
+        public void ClearNetworkErrorEvent() => NetworkErrorEvent = null;
+        public void ClearNetworkReceiveEvent() => NetworkReceiveEvent = null;
+        public void ClearNetworkReceiveUnconnectedEvent() => NetworkReceiveUnconnectedEvent = null;
+        public void ClearNetworkLatencyUpdateEvent() => NetworkLatencyUpdateEvent = null;
+        public void ClearConnectionRequestEvent() => ConnectionRequestEvent = null;
+        public void ClearDeliveryEvent() => DeliveryEvent = null;
+        public void ClearNtpResponseEvent() => NtpResponseEvent = null;
+        public void ClearPeerAddressChangedEvent() => PeerAddressChangedEvent = null;
 
-        public void ClearPeerDisconnectedEvent()
-        {
-            PeerDisconnectedEvent = null;
-        }
+        void INetEventListener.OnPeerConnected(NetPeer peer) =>
+            PeerConnectedEvent?.Invoke(peer);
 
-        public void ClearNetworkErrorEvent()
-        {
-            NetworkErrorEvent = null;
-        }
+        void INetEventListener.OnPeerDisconnected(NetPeer peer, DisconnectInfo disconnectInfo) =>
+            PeerDisconnectedEvent?.Invoke(peer, disconnectInfo);
 
-        public void ClearNetworkReceiveEvent()
-        {
-            NetworkReceiveEvent = null;
-        }
+        void INetEventListener.OnNetworkError(IPEndPoint endPoint, SocketError socketErrorCode) =>
+            NetworkErrorEvent?.Invoke(endPoint, socketErrorCode);
 
-        public void ClearNetworkReceiveUnconnectedEvent()
-        {
-            NetworkReceiveUnconnectedEvent = null;
-        }
+        void INetEventListener.OnNetworkReceive(NetPeer peer, NetPacketReader reader, byte channelNumber, DeliveryMethod deliveryMethod) =>
+            NetworkReceiveEvent?.Invoke(peer, reader, channelNumber, deliveryMethod);
 
-        public void ClearNetworkLatencyUpdateEvent()
-        {
-            NetworkLatencyUpdateEvent = null;
-        }
+        void INetEventListener.OnNetworkReceiveUnconnected(IPEndPoint remoteEndPoint, NetPacketReader reader, UnconnectedMessageType messageType) =>
+            NetworkReceiveUnconnectedEvent?.Invoke(remoteEndPoint, reader, messageType);
 
-        public void ClearConnectionRequestEvent()
-        {
-            ConnectionRequestEvent = null;
-        }
+        void INetEventListener.OnNetworkLatencyUpdate(NetPeer peer, int latency) =>
+            NetworkLatencyUpdateEvent?.Invoke(peer, latency);
 
-        public void ClearDeliveryEvent()
-        {
-            DeliveryEvent = null;
-        }
+        void INetEventListener.OnConnectionRequest(ConnectionRequest request) =>
+            ConnectionRequestEvent?.Invoke(request);
 
-        public void ClearNtpResponseEvent()
-        {
-            NtpResponseEvent = null;
-        }
+        void INetEventListener.OnMessageDelivered(NetPeer peer, object userData) =>
+            DeliveryEvent?.Invoke(peer, userData);
 
-        public void ClearPeerAddressChangedEvent()
-        {
-            PeerAddressChangedEvent = null;
-        }
+        void INetEventListener.OnNtpResponse(NtpPacket packet) =>
+            NtpResponseEvent?.Invoke(packet);
 
-        void INetEventListener.OnPeerConnected(NetPeer peer)
-        {
-            if (PeerConnectedEvent != null)
-                PeerConnectedEvent(peer);
-        }
+        void INetEventListener.OnPeerAddressChanged(NetPeer peer, IPEndPoint previousAddress) =>
+            PeerAddressChangedEvent?.Invoke(peer, previousAddress);
+    }
 
-        void INetEventListener.OnPeerDisconnected(NetPeer peer, DisconnectInfo disconnectInfo)
-        {
-            if (PeerDisconnectedEvent != null)
-                PeerDisconnectedEvent(peer, disconnectInfo);
-        }
+        /// <summary>
+    /// Simple event based listener for simple setups and benchmarks
+    /// </summary>
+    public class EventBasedLiteNetListener : ILiteNetEventListener
+    {
+        public delegate void OnPeerConnected(LiteNetPeer peer);
+        public delegate void OnPeerDisconnected(LiteNetPeer peer, DisconnectInfo disconnectInfo);
+        public delegate void OnNetworkError(IPEndPoint endPoint, SocketError socketError);
+        public delegate void OnNetworkReceive(LiteNetPeer peer, NetPacketReader reader, DeliveryMethod deliveryMethod);
+        public delegate void OnNetworkReceiveUnconnected(IPEndPoint remoteEndPoint, NetPacketReader reader, UnconnectedMessageType messageType);
+        public delegate void OnNetworkLatencyUpdate(LiteNetPeer peer, int latency);
+        public delegate void OnConnectionRequest(ConnectionRequest request);
+        public delegate void OnDeliveryEvent(LiteNetPeer peer, object userData);
+        public delegate void OnPeerAddressChangedEvent(LiteNetPeer peer, IPEndPoint previousAddress);
 
-        void INetEventListener.OnNetworkError(IPEndPoint endPoint, SocketError socketErrorCode)
-        {
-            if (NetworkErrorEvent != null)
-                NetworkErrorEvent(endPoint, socketErrorCode);
-        }
+        public event OnPeerConnected PeerConnectedEvent;
+        public event OnPeerDisconnected PeerDisconnectedEvent;
+        public event OnNetworkError NetworkErrorEvent;
+        public event OnNetworkReceive NetworkReceiveEvent;
+        public event OnNetworkReceiveUnconnected NetworkReceiveUnconnectedEvent;
+        public event OnNetworkLatencyUpdate NetworkLatencyUpdateEvent;
+        public event OnConnectionRequest ConnectionRequestEvent;
+        public event OnDeliveryEvent DeliveryEvent;
+        public event OnPeerAddressChangedEvent PeerAddressChangedEvent;
 
-        void INetEventListener.OnNetworkReceive(NetPeer peer, NetPacketReader reader, byte channelNumber, DeliveryMethod deliveryMethod)
-        {
-            if (NetworkReceiveEvent != null)
-                NetworkReceiveEvent(peer, reader, channelNumber, deliveryMethod);
-        }
+        public void ClearPeerConnectedEvent() =>  PeerConnectedEvent = null;
+        public void ClearPeerDisconnectedEvent() => PeerDisconnectedEvent = null;
+        public void ClearNetworkErrorEvent() => NetworkErrorEvent = null;
+        public void ClearNetworkReceiveEvent() => NetworkReceiveEvent = null;
+        public void ClearNetworkReceiveUnconnectedEvent() => NetworkReceiveUnconnectedEvent = null;
+        public void ClearNetworkLatencyUpdateEvent() => NetworkLatencyUpdateEvent = null;
+        public void ClearConnectionRequestEvent() => ConnectionRequestEvent = null;
+        public void ClearDeliveryEvent() => DeliveryEvent = null;
+        public void ClearPeerAddressChangedEvent() => PeerAddressChangedEvent = null;
 
-        void INetEventListener.OnNetworkReceiveUnconnected(IPEndPoint remoteEndPoint, NetPacketReader reader, UnconnectedMessageType messageType)
-        {
-            if (NetworkReceiveUnconnectedEvent != null)
-                NetworkReceiveUnconnectedEvent(remoteEndPoint, reader, messageType);
-        }
+        void ILiteNetEventListener.OnPeerConnected(LiteNetPeer peer) =>
+            PeerConnectedEvent?.Invoke(peer);
 
-        void INetEventListener.OnNetworkLatencyUpdate(NetPeer peer, int latency)
-        {
-            if (NetworkLatencyUpdateEvent != null)
-                NetworkLatencyUpdateEvent(peer, latency);
-        }
+        void ILiteNetEventListener.OnPeerDisconnected(LiteNetPeer peer, DisconnectInfo disconnectInfo) =>
+            PeerDisconnectedEvent?.Invoke(peer, disconnectInfo);
 
-        void INetEventListener.OnConnectionRequest(ConnectionRequest request)
-        {
-            if (ConnectionRequestEvent != null)
-                ConnectionRequestEvent(request);
-        }
+        void ILiteNetEventListener.OnNetworkError(IPEndPoint endPoint, SocketError socketErrorCode) =>
+            NetworkErrorEvent?.Invoke(endPoint, socketErrorCode);
 
-        void IDeliveryEventListener.OnMessageDelivered(NetPeer peer, object userData)
-        {
-            if (DeliveryEvent != null)
-                DeliveryEvent(peer, userData);
-        }
+        void ILiteNetEventListener.OnNetworkReceive(LiteNetPeer peer, NetPacketReader reader, DeliveryMethod deliveryMethod) =>
+            NetworkReceiveEvent?.Invoke(peer, reader, deliveryMethod);
 
-        void INtpEventListener.OnNtpResponse(NtpPacket packet)
-        {
-            if (NtpResponseEvent != null)
-                NtpResponseEvent(packet);
-        }
+        void ILiteNetEventListener.OnNetworkReceiveUnconnected(IPEndPoint remoteEndPoint, NetPacketReader reader, UnconnectedMessageType messageType) =>
+            NetworkReceiveUnconnectedEvent?.Invoke(remoteEndPoint, reader, messageType);
 
-        void IPeerAddressChangedListener.OnPeerAddressChanged(NetPeer peer, IPEndPoint previousAddress)
-        {
-            if (PeerAddressChangedEvent != null)
-                PeerAddressChangedEvent(peer, previousAddress);
-        }
+        void ILiteNetEventListener.OnNetworkLatencyUpdate(LiteNetPeer peer, int latency) =>
+            NetworkLatencyUpdateEvent?.Invoke(peer, latency);
+
+        void ILiteNetEventListener.OnConnectionRequest(ConnectionRequest request) =>
+            ConnectionRequestEvent?.Invoke(request);
+
+        void ILiteNetEventListener.OnMessageDelivered(LiteNetPeer peer, object userData) =>
+            DeliveryEvent?.Invoke(peer, userData);
+
+        void ILiteNetEventListener.OnPeerAddressChanged(LiteNetPeer peer, IPEndPoint previousAddress) =>
+            PeerAddressChangedEvent?.Invoke(peer, previousAddress);
     }
 }
