@@ -22,10 +22,8 @@ namespace LiteNetLib
             PeerId = localId;
         }
 
-        public static int GetProtocolId(NetPacket packet)
-        {
-            return BitConverter.ToInt32(packet.RawData, 1);
-        }
+        public static int GetProtocolId(NetPacket packet) =>
+            BitConverter.ToInt32(packet.RawData, 1);
 
         public static NetConnectRequestPacket FromData(NetPacket packet)
         {
@@ -53,23 +51,6 @@ namespace LiteNetLib
             return new NetConnectRequestPacket(connectionTime, packet.ConnectionNumber, peerId, addressBytes, reader);
         }
 
-        public static NetPacket Make(NetDataWriter connectData, SocketAddress addressBytes, long connectTime, int localId)
-        {
-            //Make initial packet
-            var packet = new NetPacket(PacketProperty.ConnectRequest, connectData.Length+addressBytes.Size);
-
-            //Add data
-            FastBitConverter.GetBytes(packet.RawData, 1, NetConstants.ProtocolId);
-            FastBitConverter.GetBytes(packet.RawData, 5, connectTime);
-            FastBitConverter.GetBytes(packet.RawData, 13, localId);
-            packet.RawData[HeaderSize-1] = (byte)addressBytes.Size;
-            for (int i = 0; i < addressBytes.Size; i++)
-                packet.RawData[HeaderSize + i] = addressBytes[i];
-            Buffer.BlockCopy(connectData.Data, 0, packet.RawData, HeaderSize + addressBytes.Size, connectData.Length);
-            return packet;
-        }
-
-#if LITENETLIB_SPANS || NETCOREAPP2_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NETCOREAPP2_1 || NETCOREAPP3_1 || NET5_0 || NETSTANDARD2_1
         public static NetPacket Make(ReadOnlySpan<byte> connectData, SocketAddress addressBytes, long connectTime, int localId)
         {
             //Make initial packet
@@ -85,7 +66,6 @@ namespace LiteNetLib
             connectData.CopyTo(packet.RawData.AsSpan(HeaderSize + addressBytes.Size));
             return packet;
         }
-#endif
     }
 
     internal sealed class NetConnectAcceptPacket
@@ -137,8 +117,8 @@ namespace LiteNetLib
             FastBitConverter.GetBytes(packet.RawData, 11, localPeerId);
             return packet;
         }
-        
-        public static NetPacket MakeNetworkChanged(NetPeer peer)
+
+        public static NetPacket MakeNetworkChanged(LiteNetPeer peer)
         {
             var packet = new NetPacket(PacketProperty.PeerNotFound, Size-1);
             FastBitConverter.GetBytes(packet.RawData, 1, peer.ConnectTime);
