@@ -128,12 +128,10 @@ namespace LiteNetLib
             byte[] pinnedBuffer,
             int len,
             byte[] socketAddress,
-            ref int socketAddressSize)
-        {
-            return UnixMode
+            ref int socketAddressSize) =>
+            UnixMode
                 ? UnixSock.recvfrom(socketHandle, pinnedBuffer, len, 0, socketAddress, ref socketAddressSize)
                 : WinSock.recvfrom(socketHandle, pinnedBuffer, len, 0, socketAddress, ref socketAddressSize);
-        }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static unsafe int SendTo(
@@ -141,12 +139,10 @@ namespace LiteNetLib
             byte* pinnedBuffer,
             int len,
             byte[] socketAddress,
-            int socketAddressSize)
-        {
-            return UnixMode
+            int socketAddressSize) =>
+            UnixMode
                 ? UnixSock.sendto(socketHandle, pinnedBuffer, len, 0, socketAddress, socketAddressSize)
                 : WinSock.sendto(socketHandle, pinnedBuffer, len, 0, socketAddress, socketAddressSize);
-        }
 
         public static SocketError GetSocketError()
         {
@@ -169,11 +165,9 @@ namespace LiteNetLib
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static short GetNativeAddressFamily(IPEndPoint remoteEndPoint)
-        {
-            return UnixMode
+        public static short GetNativeAddressFamily(IPEndPoint remoteEndPoint) =>
+            UnixMode
                 ? (short)(remoteEndPoint.AddressFamily == AddressFamily.InterNetwork ? AF_INET : AF_INET6)
                 : (short)remoteEndPoint.AddressFamily;
-        }
     }
 }
