@@ -35,35 +35,23 @@ namespace LiteNetLib
             Interlocked.Exchange(ref _packetLoss, 0);
         }
 
-        public void IncrementPacketsSent()
-        {
+        public void IncrementPacketsSent() =>
             Interlocked.Increment(ref _packetsSent);
-        }
 
-        public void IncrementPacketsReceived()
-        {
+        public void IncrementPacketsReceived() =>
             Interlocked.Increment(ref _packetsReceived);
-        }
 
-        public void AddBytesSent(long bytesSent)
-        {
+        public void AddBytesSent(long bytesSent) =>
             Interlocked.Add(ref _bytesSent, bytesSent);
-        }
 
-        public void AddBytesReceived(long bytesReceived)
-        {
+        public void AddBytesReceived(long bytesReceived) =>
             Interlocked.Add(ref _bytesReceived, bytesReceived);
-        }
 
-        public void IncrementPacketLoss()
-        {
+        public void IncrementPacketLoss() =>
             Interlocked.Increment(ref _packetLoss);
-        }
 
-        public void AddPacketLoss(long packetLoss)
-        {
+        public void AddPacketLoss(long packetLoss) =>
             Interlocked.Add(ref _packetLoss, packetLoss);
-        }
 
         public override string ToString()
         {
